[FIX] 진행 중인 소원 관련 API 경로 변경
