+++ src/main/java/com/sopterm/makeawish/controller/WishController.java
@@ -2,6 +2,7 @@ package com.sopterm.makeawish.controller;
 
 import com.sopterm.makeawish.common.ApiResponse;
 import com.sopterm.makeawish.domain.user.InternalMemberDetails;
+import com.sopterm.makeawish.dto.user.UserWishUpdateRequestDTO;
 import com.sopterm.makeawish.dto.wish.WishIdRequestDTO;
 import com.sopterm.makeawish.dto.wish.WishRequestDTO;
 import com.sopterm.makeawish.service.WishService;
@@ -20,6 +21,8 @@ import java.net.URI;
 import java.nio.file.AccessDeniedException;
 
 import static com.sopterm.makeawish.common.ApiResponse.*;
+import static com.sopterm.makeawish.common.message.ErrorMessage.*;
+import static com.sopterm.makeawish.common.message.ErrorMessage.NO_WISH;
 import static com.sopterm.makeawish.common.message.SuccessMessage.*;
 import static java.util.Objects.nonNull;
 import static org.springframework.http.HttpStatus.*;
@@ -93,6 +96,27 @@ public class WishController {
 	) {
 		wishService.deleteWishes(memberDetails.getId(), requestDTO);
 		return ResponseEntity.ok(success(SUCCESS_DELETE_WISHES.getMessage()));
+	}
+
+	@Operation(summary = "진행 중인 소원 수정", description = "수정되지 않은 정보는 null 또는 원래의 정보 그대로 전달")
+	@PutMapping("/main")
+	public ResponseEntity<ApiResponse> updateUserMainWish(
+		@Parameter(hidden = true) @AuthenticationPrincipal InternalMemberDetails memberDetails,
+		@RequestBody UserWishUpdateRequestDTO requestDTO
+	) {
+		val wish = wishService.updateUserMainWish(memberDetails.getId(), requestDTO);
+		return ResponseEntity.ok(ApiResponse.success(SUCCESS_UPDATE_USER_INFO.getMessage(), wish));
+	}
+
+	@Operation(summary = "진행 중인 소원 정보 조회")
+	@GetMapping("/main")
+	public ResponseEntity<ApiResponse> findUserMainWish(
+		@Parameter(hidden = true) @AuthenticationPrincipal InternalMemberDetails memberDetails
+	) {
+		val response = wishService.findUserMainWish(memberDetails.getId());
+		return nonNull(response)
+			? ResponseEntity.ok(ApiResponse.success(SUCCESS_GET_USER_INFO.getMessage(), response))
+			: ResponseEntity.ok(ApiResponse.fail(EXPIRED_BIRTHDAY_WISH.getMessage()));
 	}
 
 	private URI getURI(Long id) {
