+++ src/main/java/com/sopterm/makeawish/service/UserService.java
@@ -1,21 +1,15 @@
 package com.sopterm.makeawish.service;
 
-import static com.sopterm.makeawish.common.Util.*;
 import static com.sopterm.makeawish.common.message.ErrorMessage.*;
-import static com.sopterm.makeawish.domain.wish.WishStatus.*;
 import static java.util.Objects.*;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.sopterm.makeawish.domain.user.User;
-import com.sopterm.makeawish.domain.wish.Wish;
 import com.sopterm.makeawish.dto.user.UserAccountResponseDTO;
 import com.sopterm.makeawish.dto.user.UserAccountUpdateRequestDTO;
-import com.sopterm.makeawish.dto.user.UserWishUpdateResponseDTO;
-import com.sopterm.makeawish.dto.user.UserWishUpdateRequestDTO;
 import com.sopterm.makeawish.repository.UserRepository;
-import com.sopterm.makeawish.repository.wish.WishRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -27,34 +21,6 @@ import lombok.val;
 public class UserService {
 
 	private final UserRepository userRepository;
-	private final WishRepository wishRepository;
-
-	@Transactional
-	public UserWishUpdateResponseDTO updateUserMainWish(Long userId, UserWishUpdateRequestDTO request) {
-		val wisher = getUser(userId);
-		val wish =  getUserMainWish(wisher);
-
-		val status = wish.getStatus(0);
-		if (status.equals(END)) {
-			throw new IllegalArgumentException(NOT_CURRENT_WISH.getMessage());
-		}
-		if (status.equals(BEFORE)) {
-			val startDate = nonNull(request.startDate()) ? convertToDate(request.startDate()) : null;
-			val endDate = nonNull(request.endDate()) ? convertToDate(request.endDate()) : null;
-			wish.updateTerm(startDate, endDate);
-		}
-		if (status.equals(BEFORE) || status.equals(WHILE)) {
-			wisher.updateProfile(request.name(), request.bankName(), request.account(), request.phone());
-		}
-
-		return UserWishUpdateResponseDTO.of(wisher, wish);
-	}
-
-	public UserWishUpdateResponseDTO findUserMainWish(Long userId) {
-		val wisher = getUser(userId);
-		val wish =  getUserMainWish(wisher);
-		return UserWishUpdateResponseDTO.of(wisher, wish);
-	}
 
 	public UserAccountResponseDTO getUserAccount(Long userId) {
 		val wisher = getUser(userId);
@@ -77,11 +43,5 @@ public class UserService {
 	private User getUser(Long userId) {
 		return userRepository.findById(userId)
 			.orElseThrow(() -> new EntityNotFoundException(INVALID_USER.getMessage()));
-	}
-
-	private Wish getUserMainWish(User user) {
-		return   wishRepository
-			.findMainWish(user, 0)
-			.orElseThrow(() -> new EntityNotFoundException(NO_WISH.getMessage()));
 	}
 }
