+++ src/main/java/com/sopterm/makeawish/service/WishService.java
@@ -2,6 +2,7 @@ package com.sopterm.makeawish.service;
 
 import static com.sopterm.makeawish.common.Util.*;
 import static com.sopterm.makeawish.common.message.ErrorMessage.*;
+import static com.sopterm.makeawish.domain.wish.WishStatus.*;
 import static java.util.Objects.*;
 
 import java.io.IOException;
@@ -10,6 +11,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import com.sopterm.makeawish.domain.wish.WishStatus;
+import com.sopterm.makeawish.dto.user.UserWishUpdateRequestDTO;
+import com.sopterm.makeawish.dto.user.UserWishUpdateResponseDTO;
 import com.sopterm.makeawish.dto.wish.*;
 
 import org.jsoup.Jsoup;
@@ -97,6 +100,34 @@ public class WishService {
 		return WishesResponseDTO.of(wishes);
 	}
 
+
+	@Transactional
+	public UserWishUpdateResponseDTO updateUserMainWish(Long userId, UserWishUpdateRequestDTO request) {
+		val wisher = getUser(userId);
+		val wish =  getUserMainWish(wisher);
+
+		val status = wish.getStatus(0);
+		if (status.equals(END)) {
+			throw new IllegalArgumentException(NOT_CURRENT_WISH.getMessage());
+		}
+		if (status.equals(BEFORE)) {
+			val startDate = nonNull(request.startDate()) ? convertToDate(request.startDate()) : null;
+			val endDate = nonNull(request.endDate()) ? convertToDate(request.endDate()) : null;
+			wish.updateTerm(startDate, endDate);
+		}
+		if (status.equals(BEFORE) || status.equals(WHILE)) {
+			wisher.updateProfile(request.name(), request.bankName(), request.account(), request.phone());
+		}
+
+		return UserWishUpdateResponseDTO.of(wisher, wish);
+	}
+
+	public UserWishUpdateResponseDTO findUserMainWish(Long userId) {
+		val wisher = getUser(userId);
+		val wish =  getUserMainWish(wisher);
+		return UserWishUpdateResponseDTO.of(wisher, wish);
+	}
+
 	private User getUser(Long userId) {
 		return userRepository.findById(userId)
 				.orElseThrow(() -> new EntityNotFoundException(INVALID_USER.getMessage()));
@@ -124,5 +155,11 @@ public class WishService {
 	private boolean isDeletable(User user, Long wishId) {
 		val wish = getWish(wishId);
 		return wish.getWisher().equals(user) && wish.getStatus(EXPIRY_DAY).equals(WishStatus.END);
+	}
+
+	private Wish getUserMainWish(User user) {
+		return   wishRepository
+			.findMainWish(user, 0)
+			.orElseThrow(() -> new EntityNotFoundException(NO_WISH.getMessage()));
 	}
 }
